Renamed file structure scripts in order to reduce name shadowing
