+++ conf/script/src/file_structure/__init__.py
@@ -1,7 +1,7 @@
-from .build_dir import *
-from .build_system import *
-from .build_target import *
-from .cmd_cli import *
-from .conf_dir import *
-from .file_system import *
-from .root_dir import *
+from .fs_build_dir import *
+from .fs_build_system import *
+from .fs_build_target import *
+from .fs_cmd_cli import *
+from .fs_conf_dir import *
+from .fs_file_system import *
+from .fs_root_dir import *
