+++ conf/script/src/file_structure/fs_build_system.py
@@ -4,7 +4,7 @@ __all__ = ['SYSTEM_STR',
            'BUILD_SYSTEM_CONF_FILE_NAME']
 
 from ._type_alias import *
-from .build_dir import *
+from .fs_build_dir import *
 
 SYSTEM_STR: TAlias_Name = 'system'
 BUILD_SYSTEM_NAME: TAlias_Name = 'meson'
