+++ conf/script/src/file_structure/fs_conf_dir.py
@@ -3,8 +3,8 @@ __all__ = ['CONF_DIR_NAME',
            'CONF_MESON_MACHINE_FILES_DIR_NAME']
 
 from ._type_alias import *
-from .build_dir import *
-from .build_system import *
+from .fs_build_dir import *
+from .fs_build_system import *
 
 CONF_DIR_NAME: TAlias_Name = 'conf'
 CONF_BUILD_SYSTEM_DIR_NAME: TAlias_Name = f'{BUILD_DIR_NAME}-{SYSTEM_STR}'
